requirements.txt
sqlalchemy==2.0.40
alembic==1.15.2
PyMySQL==1.1.1
python-dotenv==1.0.1
requests==2.32.3
pandas==2.2.3
cryptography==44.0.2
